El merge va bien

+++ Models/Transaccion.cs
@@ -3,7 +3,7 @@ namespace APIAA.Models;
 public class Transaccion
 {
     public int Id { get; set; }
-    public decimal Unidades { get; set;}
+    public double Unidades { get; set;}
     public DateTime Fecha { get; set;}
     public int VideojuegoId {get; set;}
